# api/router/handlers/search_handlers.go
package handlers

import (
	"net/http"
)

// SearchTrafficHandler is a placeholder for searching traffic logs.
// @Summary Search traffic logs
// @Description (Not Implemented Yet) Searches through the `http_traffic_log` table based on a query string. Searchable fields could include URLs, headers, and bodies.
// @Tags Search
// @Produce json
// @Param q query string true "Search query term"
// @Param target_id query int false "Optional ID of the target to scope the search to"
// @Param page query int false "Page number for pagination" default(1)
// @Param limit query int false "Number of items per page" default(50)
// @Success 501 {object} models.ErrorResponse "Not Implemented Yet"
// @Failure 400 {object} models.ErrorResponse "Missing query parameter 'q'"
// @Router /search/traffic [get]
func SearchTrafficHandler(w http.ResponseWriter, r *http.Request) { notImplementedHandler(w, r) }